照片墙接口 修复文件夹不存在bug

+++ router_handler/article.js
@@ -27,15 +27,27 @@ exports.artPost = (req, res) => {
                     const fileUrl = baseURL + '/images/article/articleID-' + artidDoc.artid + '/'
                     const newfilename = 'article-Cover' + '.' + extname //封面命名
                     const CoverUrl = fileUrl + newfilename //封面URL
-                    fs.mkdir('./public/images/article/articleID-' + artidDoc.artid + '/', () => { }) //创建文章图片文件夹
-                    fs.renameSync('./uploads/' + file.filename, './public/images/article/articleID-' + artidDoc.artid + '/' + newfilename) //重命名图片
-                    const userDoc = await userModule.findOne({ 'uid': user.uid }) //查询作者
-                    const addArt = await articleModule.create({
-                        'artContent': info.articleContent, 'artId': artidDoc.artid, 'artImages': CoverUrl, 'artTitle': info.articleTitle, 'author': userDoc, 'date': info.articleDate
-                    }) //添加文章
-                    const resetId = await imageIdModule.findOneAndUpdate({ name: 'image' }, { $set:{imageId: 0} }, { new: true })//reset image id
-                    const userChan = await userModule.findOneAndUpdate({ 'uid': user.uid }, { $push: { 'article': artidDoc.artid } })//将文章id添加到相应用户
-                    res.back(300, '文章发布成功!')
+                    const exists = './public/images/article/articleID-' + artidDoc.artid + '/'
+                    if (fs.existsSync(exists)) {
+                        fs.renameSync('./uploads/' + file.filename, './public/images/article/articleID-' + artidDoc.artid + '/' + newfilename) //重命名图片
+                        const userDoc = await userModule.findOne({ 'uid': user.uid }) //查询作者
+                        const addArt = await articleModule.create({
+                            'artContent': info.articleContent, 'artId': artidDoc.artid, 'artImages': CoverUrl, 'artTitle': info.articleTitle, 'author': userDoc, 'date': info.articleDate
+                        }) //添加文章
+                        const resetId = await imageIdModule.findOneAndUpdate({ name: 'image' }, { $set: { imageId: 0 } }, { new: true })//reset image id
+                        const userChan = await userModule.findOneAndUpdate({ 'uid': user.uid }, { $push: { 'article': artidDoc.artid } })//将文章id添加到相应用户
+                        res.back(300, '文章发布成功!')
+                    } else {
+                        fs.mkdirSync('./public/images/article/articleID-' + artidDoc.artid + '/') //创建文章图片文件夹
+                        fs.renameSync('./uploads/' + file.filename, './public/images/article/articleID-' + artidDoc.artid + '/' + newfilename) //重命名图片
+                        const userDoc = await userModule.findOne({ 'uid': user.uid }) //查询作者
+                        const addArt = await articleModule.create({
+                            'artContent': info.articleContent, 'artId': artidDoc.artid, 'artImages': CoverUrl, 'artTitle': info.articleTitle, 'author': userDoc, 'date': info.articleDate
+                        }) //添加文章
+                        const resetId = await imageIdModule.findOneAndUpdate({ name: 'image' }, { $set: { imageId: 0 } }, { new: true })//reset image id
+                        const userChan = await userModule.findOneAndUpdate({ 'uid': user.uid }, { $push: { 'article': artidDoc.artid } })//将文章id添加到相应用户
+                        res.back(300, '文章发布成功!')
+                    }
                 }
             }
         } else {
@@ -58,7 +70,6 @@ exports.artDelete = (req, res) => {
     async function handler() {
         const uid = req.auth.uid
         const artId = req.body.artId
-        const artidDoc = await articleIdModule.findOneAndUpdate({ name: 'article' }, { $inc: { artid: -1 } }, { new: true })
         const artDel = await articleModule.deleteOne({ 'artId': artId })
         const userArtDel = await userModule.findOneAndUpdate({ 'uid': uid }, { $pull: { 'article': Number(artId) } })
         const artCmtDel = await commentModule.deleteMany({ 'artId': artId })
@@ -76,10 +87,19 @@ exports.artImg = (req, res) => {
         const artidIncDoc = await articleIdModule.findOneAndUpdate({ name: 'article' }, { $inc: { artid: 1 } }, { new: true })
         const fileUrl = baseURL + '/images/article/articleID-' + artidIncDoc.artid + '/'
         const newImagesUrl = fileUrl + newfilename
-        fs.mkdirSync('./public/images/article/articleID-' + artidIncDoc.artid + '/', () => { }) //创建文章图片文件夹
-        fs.renameSync('./uploads/' + files.filename, './public/images/article/articleID-' + artidIncDoc.artid + '/' + newfilename) //写入图片
-        const artiddecDoc = await articleIdModule.findOneAndUpdate({ name: 'article' }, { $inc: { artid: -1 } }, { new: true })
-        res.send(newImagesUrl)
+        const exists = './public/images/article/articleID-' + artidIncDoc.artid + '/'
+        if (fs.existsSync(exists)) {
+            fs.renameSync('./uploads/' + files.filename, './public/images/article/articleID-' + artidIncDoc.artid + '/' + newfilename)
+            const artiddecDoc = await articleIdModule.findOneAndUpdate({ name: 'article' }, { $inc: { artid: -1 } }, { new: true })
+            res.send(newImagesUrl)
+        }
+        else {
+            fs.mkdirSync('./public/images/article/articleID-' + artidIncDoc.artid + '/') //创建文章图片文件夹
+            fs.renameSync('./uploads/' + files.filename, './public/images/article/articleID-' + artidIncDoc.artid + '/' + newfilename) //写入图片
+            const artiddecDoc = await articleIdModule.findOneAndUpdate({ name: 'article' }, { $inc: { artid: -1 } }, { new: true })
+            res.send(newImagesUrl)
+        }
+
     }
     handler()
 }
